VIC-2467 Disable QCMAP_ConnectionManager service
QCMAP_ConnectionManager seems to exist for the purpose of managing Access Point configuration.
We don't want to run this.

+++ poky/meta-qti-data-prop/recipes/data/data_git.bb
@@ -124,7 +124,7 @@ do_install_append_msm() {
         if ${@bb.utils.contains('DISTRO_FEATURES', 'systemd', 'true', 'false', d)}; then
           install -d ${D}${systemd_unitdir}/system/
           #QCMAP service
-          if (test "x${BASEPRODUCT}" != "xdrone"); then
+          if (test "x${BASEPRODUCT}" != "xdrone" && test "x${BASEPRODUCT}" != "xrobot"); then
             rm -rf ${D}${sysconfdir}/init.d/start_QCMAP_ConnectionManager_le
             install -m 0644 ${WORKDIR}/QCMAP_ConnectionManagerd.service -D ${D}${systemd_unitdir}/system/QCMAP_ConnectionManagerd.service
             install -d ${D}${systemd_unitdir}/system/multi-user.target.wants/
